build_library: check for EROOT in preserved environment
This is a long standing bug that bites us when ebuilds try to set EROOT
and related variables in EAPIs < 3. When binary packages are installed
to a different ROOT than they were built for but EROOT remains many
strange and hard to debug issues may occur. So far we've gotten by
because this wasn't a common idiom and only glibc and binutils did it.

Well, now it is a common idiom, for example `eutils.eclass` now has:

    # Initialized prefix variables for EAPI<3.
    _eutils_eprefix_init() {
        has "${EAPI:-0}" 0 1 2 && : ${ED:=${D}} ${EPREFIX:=} ${EROOT:=${ROOT}}
    }

This reason alone is why I haven't updated much in portage-stable lately
because a spectacular number of things will now be impacted by this.
Before fixing the bug in portage properly here is a test to fail loudly
if the fix doesn't work, errors would be hard to spot otherwise.